Completes bridge pattern
